Updated to use RandomNumberGenerator
Code now uses RandomNumberGenerator instead of RNGCryptoServiceProvider. 
Ranges for accounts have been changed - in particular Sales cannot be less than 3/4 of Purchases.
First account to be initialized is randomized. Accounts have an initial 1/3 chance to be given now.
Added a function to check if an AccountVal can be solved with its own substitutes.