Main algorithm updated. Start button now works.
The algorithm now sets the first randomly selected account to needs solving. Additionally, it now works (previous update resulted in an infinite loop due to not removing elements from the tempArray).
StartButton now runs a method which initializes values and displays the information needed to solve the generated problem(s).