Questions added, submitting answers now works
Program now generates questions and contains a submission button and textbox. Submission is checked through using the 'Enter' or 'Return' key or by pressing the 'Submit' Button. 
Added Dictionaries that corresponds the submission textbox and the submit button to the account needing to be solved for.
Also the input box flashes red when the answer is wrong.