Removed some randomization of a var's solvability
Removed the 1/3 chance of a variable being set as 'given' in the question set. This makes it so the chance of a variable being given is solely determined by the random order selection. This should make the algorithm have a lower chance of having 0 unsolved variables (I am hoping it is a 0% chance though).