Main algorithm updated
New algorithm uses the IsSolvable methods of the AccountVal class. New method is called 'setSolveStatesAlternative' and results in 5-6 accounts being asked for in the question set now (previously it was 2-5 with it usually appearing at 3).
Fixed the warning box so now it is read-only. Input box caret brush explicitly set to black now (previously it would show up as an aqua).
Fixed IsSolvable, UsableSubstitute, IsSolvableRecursive functions as the one of the LINQ statements acted opposite of the desired behavior.
Fixed the calculation for Cost of Merchandise Available for Sale. Previously it was based on '_purchases' as opposed to '_netPurchases'